Refactor ASM usage from rewiring package into code package.
