+++ djvm/src/main/kotlin/net/corda/djvm/code/SandboxClassRemapper.kt
@@ -1,4 +1,4 @@
-package net.corda.djvm.rewiring
+package net.corda.djvm.code
 
 import net.corda.djvm.analysis.AnalysisConfiguration
 import net.corda.djvm.analysis.AnalysisConfiguration.Companion.KOTLIN_METADATA
