+++ djvm/src/main/kotlin/net/corda/djvm/code/SewingKit.kt
@@ -1,5 +1,5 @@
 @file:JvmName("SewingKit")
-package net.corda.djvm.rewiring
+package net.corda.djvm.code
 
 import org.objectweb.asm.AnnotationVisitor
 import org.objectweb.asm.ClassVisitor
